Split out IconComponent, added comments for explanation
