+++ src/icons/templates/withFdsIconWrapper.jsx
@@ -1,35 +1,16 @@
 import React from 'react';
-import cx from 'classnames';
-import PropTypes from 'prop-types';
+import IconComponent from './IconComponent';
 
-const IconComponent = (props) => (
-  <div
-    className={cx('fds-icon', `fds-icon--${props.size}`, props.className)}
-    style={{
-      fill: props.color,
-      width: `${props.customSize}px`,
-      height: `${props.customSize}px`,
-    }}
-  >
-    {props.children}
-  </div>
-);
-
-IconComponent.defaultProps = {
-  size: 's',
-};
-
-IconComponent.propTypes = {
-  color: PropTypes.string,
-  size: PropTypes.oneOf(['xs', 's', 'm', 'l', 'xl']),
-  className: PropTypes.string,
-  customSize: PropTypes.number,
-  children: PropTypes.node,
-};
+// Brief explanation: we're customizing the component by wrapping
+// it with another component, and adding customizations to the
+// wrapper component. Please refer to `IconComponent` for
+// those customizations
 
 export const withFdsIconWrapper = (WrappedComponent) => {
-  const IconName = WrappedComponent.name;
-  IconComponent.displayName = IconName;
+  // We're wrapping the component we get with our component,
+  // but we want the wrapper to be synonyous with the icon name,
+  // for testing / identification purposes
+  IconComponent.displayName = WrappedComponent.name;
   return (
     <IconComponent>
       <WrappedComponent />
